Remove duplicate boilerplate webhook
